+++ CHANGELOG.md
@@ -4,12 +4,13 @@
 
 ###bug fixes
 *The default preferences used by Selenium are now merged into the preferences of Firefox profiles loaded from disk.
+*addExtension will now correctly add extension directories to a profile.
 
 ###API changes
 * The representation of profile files has been changed to use a HashMap instead of an association list. This ensures that destination paths are always unique.
 
 ###new features 
-*It's now possible to add entire directories to a profile in pure code.
+*It's now possible to add entire directories to a profile in pure code using addFile and addExtension.
 *new functions in Common.Profile: unionProfiles, onProfileFiles, onProfilePrefs